CM-176: ability to start distinguished onClone pipeline per target
